requirements.txt
wheel
beautifulsoup4
certifi
chardet
comtypes
decorator
idna
imageio
imageio-ffmpeg
moviepy
numpy==1.19.3
pafy
pandas
pathtools
Pillow
playsound
proglog
PyAudio
python-dateutil
python-vlc
pyttsx3
pytube
pytz
pywin32
requests
selenium
six
soupsieve
SpeechRecognition
tqdm
urllib3
watchdog
wikipedia
youtube-dl
chromedriver_autoinstaller
configparser
pywebview
youtube-search